% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{base_url_request}
\alias{base_url_request}
\title{base function to return the data for the specified activity based on the url}
\usage{
base_url_request(
  url,
  oauth_token,
  show_nchar_case_error = 135,
  simplifyVector = TRUE
)
}
\arguments{
\item{url}{a character string specifying the input url}

\item{oauth_token}{a character string specifying the authentication token}

\item{show_nchar_case_error}{an integer value specifying the number of characters to show in case of an error}

\item{simplifyVector}{a boolean. Coerce JSON arrays containing only primitives into an atomic vector (see the documentation of the jsonlite::fromJSON() function)}
}
\value{
an object of class list
}
\description{
base function to return the data for the specified activity based on the url
}
\keyword{internal}
