% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{ggplot_each_date}
\alias{ggplot_each_date}
\title{plot function for a single day (heart rate)}
\usage{
ggplot_each_date(date_intraday, date)
}
\arguments{
\item{date_intraday}{a data.table of Intraday Dates of heart rate measurements}

\item{date}{a character string specifying a Date}
}
\value{
a plot object of class ggplot2
}
\description{
plot function for a single day (heart rate)
}
\keyword{internal}
