% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{heart_rate_time_series}
\alias{heart_rate_time_series}
\title{heart rate activity time series}
\usage{
heart_rate_time_series(
  heart_rate_intraday_list,
  heart_rate = NULL,
  detail_level = "1min",
  ggplot_intraday = FALSE,
  ggplot_ncol = NULL,
  ggplot_nrow = NULL,
  verbose = FALSE
)
}
\arguments{
\item{heart_rate_intraday_list}{a named list of data.tables with the intraday heart rate data. Each element is a data.table
with columns \code{time} (character, format "YYYY-MM-DD HH:MM:SS") and \code{value} (numeric heart rate).
Names must be dates in "YYYY-MM-DD" format. This is the pre-downloaded intraday heart rate data.}

\item{heart_rate}{either NULL or a data.frame with the daily heart rate summary data}

\item{detail_level}{a character string specifying the detail level of the heart rate time series. It can be either '1min' or '1sec', for 1-minute and 1-second intervals}

\item{ggplot_intraday}{a boolean. If TRUE then the ggplot of the heart rate time series will be returned too}

\item{ggplot_ncol}{either NULL or an integer specifying the number of columns of the output ggplot}

\item{ggplot_nrow}{either NULL or an integer specifying the number of rows of the output ggplot}

\item{verbose}{a boolean. If TRUE then information will be printed out in the console}
}
\value{
an object of class list
}
\description{
heart rate activity time series
}
\examples{
\dontrun{

require(fitbitViz)

# load pre-downloaded intraday heart rate data (one data.table per day, named by date)
# heart_rate_intraday_list is a named list where each element has columns 'time' and 'value'

heart_dat <- heart_rate_time_series(
  heart_rate_intraday_list = heart_rate_intraday_list,
  heart_rate = heart_rate_daily,
  detail_level = "1min",
  ggplot_intraday = TRUE,
  verbose = TRUE
)
heart_dat$plt
heart_dat$heart_rate
heart_dat$heart_rate_intraday
}
}
