% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{meshgrids_XY_LatLon}
\alias{meshgrids_XY_LatLon}
\title{Find the image (x,y)-coordinates using meshgrids of (lat,lon)-coordinates}
\usage{
meshgrids_XY_LatLon(
  longitude,
  latitude,
  buffer_raster,
  buffer_bbox,
  distance_metric = "vincenty",
  digits = 8
)
}
\arguments{
\item{longitude}{a numeric value specifying the longitude}

\item{latitude}{a numeric value specifying the latitude}

\item{buffer_raster}{an object of class raster}

\item{buffer_bbox}{a numeric vector specifying the bounding box}

\item{distance_metric}{a character string specifying the distance metric, one of "haversine" "vincenty", "geodesic" or "cheap" (see the 'geodist::geodist()' function of the 'geodist' package)}

\item{digits}{an integer value specifying the number of digits}
}
\value{
an object of class list
}
\description{
Find the image (x,y)-coordinates using meshgrids of (lat,lon)-coordinates
}
\details{
I've used this workaround to label the rayshader 3d-plot because the 'lat', 'lon' parameters did not work properly
}
\keyword{internal}
