% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{plot_data_type}
\alias{plot_data_type}
\title{Refresh Token of an existing application}
\usage{
plot_data_type(dat_type_min, type)
}
\arguments{
\item{client_id}{a character string specifying the 'client_id' of the registered (existing) Fitbit application}

\item{client_secret}{a character string specifying the 'client_secret' of the registered (existing) Fitbit application}

\item{refresh_token}{a character string specifying the 'refresh_token' of the registered (existing) Fitbit application}
}
\value{
a named list that includes access_token, expires_in, refresh_token, scope, token_type, user_id
}
\description{
Refresh Token of an existing application
}
\details{
This plot function is used in the 'fitbit_data_type_by_date' internally to plot the 'spo2' and 'hrv' data types
}
\keyword{internal}
