% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{sleep_heatmap}
\alias{sleep_heatmap}
\title{sleep data heatmap}
\usage{
sleep_heatmap(level_data, angle_x_axis = 0)
}
\arguments{
\item{level_data}{a data.table specifying the input level data}

\item{angle_x_axis}{a float number specifying the angle of the x-axis text of the output ggplot}
}
\value{
a plot object of class ggplot2
}
\description{
sleep data heatmap
}
\keyword{internal}
