% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{split_year_in_weeks}
\alias{split_year_in_weeks}
\title{function for the weeks (including the date-from and date-to)}
\usage{
split_year_in_weeks(year)
}
\arguments{
\item{year}{an integer value specifying the year}
}
\value{
a sequence of Dates of class 'Date'
}
\description{
function for the weeks (including the date-from and date-to)
}
\references{
https://statistics.berkeley.edu/computing/faqs/dates-and-times-r
}
\keyword{internal}
