% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbit_utils.R
\name{plot_data_type}
\alias{plot_data_type}
\title{This plot function is used in the 'fitbit_data_type_by_date' internally to plot the 'spo2' and 'hrv' data types}
\usage{
plot_data_type(dat_type_min, type)
}
\description{
This plot function is used in the 'fitbit_data_type_by_date' internally to plot the 'spo2' and 'hrv' data types
}
\keyword{internal}
